database: Abort barriers on exception
The database::shutdown() and ::drain() methods are called inside the
container().invoke_on_all() and synchronize with each other via the
cross-shard _stop_barrier. If either shard throws in between all others
may get stuck waiting for the barrier to collect all arrivals.

The fix is to abort the barrier on exception thus making all the
shards sitting in shutdown or drain to bail out with exceptions too.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>