toolchain: set the PATH for the dbuild image includes ccache
ccache is added to the PATH by the `/etc/profile.d/ccache.sh` script but
currently there is an issue with the dbuild script that fails to load the
bash env. and ccache is missing from the PATH,

This commit makes sure that ccache will always be included in the PATH
regardless of the dbuild script.