expressions: avoid invalid behavior when evaluating map subscripts
When we have an expression with a null or unset subscript, i.e.,
m[?] with ? bound to a null or UNSET_VALUE, the current code which
evaluates this expression implicitly assumes that the subscript is
a valid value of the right type and deserializes it. This assumption
can cause crashes (see #10399) or bizarre errors (see #10361).

In this patch, when we see an invalid subscript, we return a nullopt
signifying that the expression could not be calculated. When this
expression is used as a filter, the considered row will silently not
match.

This patch fixes #10399 because requests with null or unset values
will no longer cause dereferencing invalid data. However,
issue #10361 remains open because when a null or unset_value *constant*
is used in a filtering expression, we expect to see an error - not
silently matching nothing.

The rationale of the fix here is that we'll need to add, in a later patch,
an expression-resolution stage which checks whether an expression after
binding - but before evaluation - has a null or unset constant subscript
(we need to verify this even if there is no data in the table). But even
if we do that, and the if() introduced in this patch into evaluation
becomess less criticial, it will still be important in the future if we
add support for non-constant subscripts (e.g., m[a] where a is another
column) - we'll still need the check introduced in this patch to just
skip rows with missing "a" instead of crashing on them.

In this patch we also introduce another test which demonstrates the
fact that we don't generate errors on null or UNSET_VALUE subscripts.
After this patch, the new test will no longer crash or cause weird
errors, but it won't cause any errors at all (just empty results),
so it is another, better, reproducer for #10361. The new test also
starts with an empty table, so it verifies that the validation we're
missing must be done before evaluating the expression - not during
the evaluation itself (which won't happen when the the table is empty).

Fixes #10399.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>