data_dictionary,sstables: add SHARED_FILESYSTEM storage type
The SHARED_FILESYSTEM type uses a shared resource which is simply a
local filesystem directory. The assumption is that this mode is now only
used in tests, which run multiple Scylla nodes on a single machine.
These machines have access to the local filesystem, so the 'dir'
parameter for storage is used to indicate the local path in which all
the sstables from shared sstables are stored.

This patch is not useful on its own unless sstables start having
globally unique names. Otherwise, file collisions are expected
in the shared directory.

Verified manually:

cassandra@cqlsh> select * from system_schema.scylla_keyspaces ;

 keyspace_name | storage_options         | storage_type
---------------+-------------------+-------------------
           foo | {'directory': '/tmp/x'} | SHARED_FILESYSTEM

[sarna@localhost scylla-ccm]$ ls -ls /tmp/x/
total 40
4 -rw-r--r--. 1 sarna sarna   66 Apr 21 17:35 me-2-big-CompressionInfo.db
4 -rw-r--r--. 1 sarna sarna   33 Apr 21 17:35 me-2-big-Data.db
4 -rw-r--r--. 1 sarna sarna   10 Apr 21 17:35 me-2-big-Digest.crc32
4 -rw-r--r--. 1 sarna sarna   16 Apr 21 17:35 me-2-big-Filter.db
4 -rw-r--r--. 1 sarna sarna    8 Apr 21 17:35 me-2-big-Index.db
4 -rw-r--r--. 1 sarna sarna  206 Apr 21 17:35 me-2-big-Scylla.db
8 -rw-r--r--. 1 sarna sarna 4581 Apr 21 17:35 me-2-big-Statistics.db
4 -rw-r--r--. 1 sarna sarna   56 Apr 21 17:35 me-2-big-Summary.db
4 -rw-r--r--. 1 sarna sarna  102 Apr 21 17:35 me-2-big-TOC.txt