service: migration_manager: check for existing group0 when doing raft-based schema operations
`raft_group_registry::is_enabled()` check is not what
migration_manager should do to check if it can use
raft for schema changes.

The group registry is enabled in raft-experimental mode, no
matter if the cluster has fully switched to using raft cluster
management or not.

Instead, it should check whether the cluster has a working
group0 and the current node is part of it. That means that
the cluster underwent a full upgrade procedure and is ready
to start using raft for schema and topology changes.

Tests: unit(dev), dtest(manual)

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>